fix(WorkflowIndexLoading: rename WorkflowLoading.jsx to WorkflowIndexLoading
